+++ src/controllers/JacobsSonsController.js
@@ -1,9 +1,11 @@
 const dbConnect = require("../database/connection");
+const { StatusCodes } = require("http-status-codes");
+const table = "jacobs_sons";
 
 module.exports = {
   async read(request, response) {
     
-    const dbResponse = await dbConnect("jacobs_sons").select("*");
-    return response.status(200).json(dbResponse);
+    const dbResponse = await dbConnect(table).select("*");
+    return response.status(StatusCodes.OK).json(dbResponse);
   },
 };