+++ src/controllers/PasswordController.js
@@ -1,11 +1,19 @@
 const dbConnect = require("../database/connection");
 const { StatusCodes } = require("http-status-codes");
-const table = "stickers_status";
+const table = "passwords";
 
 module.exports = {
   async read(request, response) {
-    
-    const dbResponse = await dbConnect(table).select("*");
+    const { id } = request.params;
+
+    const dbResponse = await dbConnect(table)
+      .select("*")
+      .where("id", id)
+      .first();
+
     return response.status(StatusCodes.OK).json(dbResponse);
   },
+
+  
+
 };