+++ src/controllers/utils/errosMessages.js
@@ -1,0 +1,8 @@
+const ErrorMessages = 
+{
+    "lackOfPassword": "It misses password",
+    "passwordSize": "Passwords must have between 8 and 25 characters",
+    "termsOfUse": "Accept Terms of Use"
+}
+
+exports.ErrorMessages = ErrorMessages